requirements.txt
# Core — all packages have pre-built wheels on PyPI
streamlit>=1.35.0,<2.0.0
openai>=1.40.0,<2.0.0
pandas>=2.0.0,<3.0.0
azure-identity>=1.16.0,<2.0.0
